+++ src/grid/gridstorage/kokkos-storage.hpp
@@ -70,10 +70,10 @@ namespace schnek {
       static constexpr size_t rank = rank_t;
 
       /// The grid index type
-      typedef Array<int, rank_t> IndexType;
+      typedef Array<ptrdiff_t, rank_t> IndexType;
 
       /// The grid range type
-      typedef Range<int, rank_t> RangeType;
+      typedef Range<ptrdiff_t, rank_t> RangeType;
 
     private:
       typedef std::function<void(const RangeType &)> UpdaterType;
@@ -146,16 +146,16 @@ namespace schnek {
       SCHNEK_INLINE const IndexType &getDims() const { return this->dims; }
 
       /// Get k-th component of the lowest coordinate in the grid (inclusive)
-      SCHNEK_INLINE int getLo(int k) const { return this->range.getLo(k); }
+      SCHNEK_INLINE ptrdiff_t getLo(size_t k) const { return this->range.getLo(k); }
 
       /// Get k-th component of the highest coordinate in the grid (inclusive)
-      SCHNEK_INLINE int getHi(int k) const { return this->range.getHi(k); }
+      SCHNEK_INLINE ptrdiff_t getHi(size_t k) const { return this->range.getHi(k); }
 
       /// Get k-th component of the dimensions of the grid `dims = high - low + 1`
-      SCHNEK_INLINE int getDims(int k) const { return this->dims[k]; }
+      SCHNEK_INLINE size_t getDims(size_t k) const { return this->dims[k]; }
 
       /// Get the length of the allocated array
-      SCHNEK_INLINE int getSize() const { return this->size; }
+      SCHNEK_INLINE size_t getSize() const { return this->size; }
 
       /**
        * @brief resizes to grid with lower indices low[0],...,low[rank-1]
