+++ testsuite/test_arrayexpression.cpp
@@ -245,8 +245,8 @@ BOOST_FIXTURE_TEST_CASE( grid_constructor, ArrayExpressionTest )
   }
 
   for (int i=0; i<N; i++) {
-    Array<int,2, ArrayBoostTestArgCheck> s1(idist_small(rGen), idist_small(rGen));
-    Array<int,2, ArrayBoostTestArgCheck> s2(idist_small(rGen), idist_small(rGen));
+    Array<ptrdiff_t,2, ArrayBoostTestArgCheck> s1(idist_small(rGen), idist_small(rGen));
+    Array<ptrdiff_t,2, ArrayBoostTestArgCheck> s2(idist_small(rGen), idist_small(rGen));
     Grid<int,2> grid(s1+s2);
 
     BOOST_CHECK_EQUAL(grid.getLo()[0], 0);
@@ -258,8 +258,8 @@ BOOST_FIXTURE_TEST_CASE( grid_constructor, ArrayExpressionTest )
   }
 
   for (int i=0; i<N; i++) {
-    Array<int,2> s1(idist_small(rGen), idist_small(rGen));
-    Array<int,2> s2(idist_small(rGen), idist_small(rGen));
+    Array<ptrdiff_t,2> s1(idist_small(rGen), idist_small(rGen));
+    Array<ptrdiff_t,2> s2(idist_small(rGen), idist_small(rGen));
     Grid<int,2> grid(-s1-s2, s1+s2);
 
     BOOST_CHECK_EQUAL(grid.getLo()[0], -s1[0]-s2[0]);
@@ -271,8 +271,8 @@ BOOST_FIXTURE_TEST_CASE( grid_constructor, ArrayExpressionTest )
   }
 
   for (int i=0; i<N; i++) {
-    Array<int,2, ArrayBoostTestArgCheck> s1(idist_small(rGen), idist_small(rGen));
-    Array<int,2, ArrayBoostTestArgCheck> s2(idist_small(rGen), idist_small(rGen));
+    Array<ptrdiff_t,2, ArrayBoostTestArgCheck> s1(idist_small(rGen), idist_small(rGen));
+    Array<ptrdiff_t,2, ArrayBoostTestArgCheck> s2(idist_small(rGen), idist_small(rGen));
     Grid<int,2> grid(-s1-s2, s1+s2);
 
     BOOST_CHECK_EQUAL(grid.getLo()[0], -s1[0]-s2[0]);
